requirements.txt
apache-airflow==2.9.0
apache-airflow-providers-google>=10.0.0
pandas>=1.4.0
openpyxl>=3.0.0
